Report Generator Tool/requirements.txt
comtypes==1.1.10
img2pdf==0.4.3
openpyxl==3.0.9
pdf2image==1.16.0
pdfplumber==0.5.28
Pillow==8.4.0
PyPDF2==1.26.0
pytesseract==0.3.8
pywin32==302
